[K1P-209] refactor : 주문 페이지 선택 사항들 컴포넌트화
