+++ frontend/sellution-app/src/shopping/layout/order/CouponSelection.jsx
@@ -1,0 +1,11 @@
+const CouponSelection = () => {
+  return (
+    <div className='mb-6 bg-white py-4 w-[90%]'>
+      <span className='block py-2 mb-2 font-bold'>할인쿠폰</span>
+      <select className='w-full p-2 border rounded'>
+        <option>신규회원 10% 할인 쿠폰</option>
+      </select>
+    </div>
+  );
+};
+export default CouponSelection;
